refactor: improving order state context management
