+++ src/main.tsx
@@ -4,11 +4,14 @@ import "./index.css";
 import { RouterProvider } from "react-router";
 import { router } from "./routes.tsx";
 import { SessionProvider } from "./web/context/session.context.tsx";
+import { OpenOrdersProvider } from "./web/context/open-orders.context.tsx";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
     <SessionProvider>
-      <RouterProvider router={router} />
+      <OpenOrdersProvider>
+        <RouterProvider router={router} />
+      </OpenOrdersProvider>
     </SessionProvider>
   </StrictMode>
 );
